Replace avx instructions with sse instructions
right now interpolation in doubles is a mess of AVX(_m256*) and SSE instructions
since feature is called sse this commit removes everything AVX for now
