+++ Cargo.toml
@@ -13,6 +13,7 @@ edition = "2018"
 
 [features]
 sys = ["speexdsp-sys"]
+sse3 = [ "speexdsp-resampler/sse3" ]
 
 [dev-dependencies]
 assert_approx_eq = "1.1.0"
