+++ resampler/src/speex.rs
@@ -833,349 +833,9 @@ fn resampler_basic_zero(
     out_sample as i32
 }
 
-mod native {
-    #[inline(always)]
-    pub fn cubic_coef(frac: f32, interp: &mut [f32]) {
-        interp[0] = -0.16666999459266663 * frac
-            + 0.16666999459266663 * frac * frac * frac;
-        interp[1] = frac + 0.5 * frac * frac - 0.5f32 * frac * frac * frac;
-        interp[3] = -0.3333300054073334 * frac + 0.5 * frac * frac
-            - 0.16666999459266663 * frac * frac * frac;
-        interp[2] =
-            (1.0f64 - interp[0] as f64 - interp[1] as f64 - interp[3] as f64)
-                as f32;
-    }
-
-    #[inline(always)]
-    pub fn interpolate_step_single(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        oversample: usize,
-        offset: usize,
-        n: usize,
-        sinc_table: &[f32],
-        frac: f32,
-    ) {
-        let mut accum: [f32; 4] = [0.; 4];
-        in_slice.iter().zip(0..n).for_each(|(&curr_in, j)| {
-            let idx = (2 + (j + 1) * oversample as usize) - offset as usize;
-            accum.iter_mut().zip(sinc_table.iter().skip(idx)).for_each(
-                |(v, &s)| {
-                    *v += curr_in * s;
-                },
-            );
-        });
-        let mut interp: [f32; 4] = [0.; 4];
-        cubic_coef(frac, &mut interp);
-        out_slice[(out_stride * out_sample) as usize] = interp
-            .iter()
-            .zip(accum.iter())
-            .map(|(&x, &y)| x * y)
-            .fold(0., |acc, x| acc + x);
-    }
-
-    #[inline(always)]
-    pub fn interpolate_step_double(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        oversample: usize,
-        offset: usize,
-        n: usize,
-        sinc_table: &[f32],
-        frac: f32,
-    ) {
-        let mut accum: [f64; 4] = [0.0; 4];
-        in_slice.iter().zip(0..n).for_each(|(&curr_in, j)| {
-            let idx = (2 + (j + 1) * oversample as usize) - offset as usize;
-            accum.iter_mut().zip(sinc_table.iter().skip(idx)).for_each(
-                |(v, &s)| {
-                    *v += (curr_in * s) as f64;
-                },
-            );
-        });
-        let mut interp: [f32; 4] = [0.; 4];
-        cubic_coef(frac, &mut interp);
-        out_slice[(out_stride * out_sample) as usize] = interp
-            .iter()
-            .zip(accum.iter())
-            .map(|(&x, &y)| x * y as f32)
-            .fold(0., |acc, x| acc + x);
-    }
-
-    #[inline(always)]
-    pub fn direct_step_single(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        n: usize,
-        sinc_table: &[f32],
-    ) {
-        let mut sum: f32 = 0.0;
-        let mut j = 0;
-        while j < n {
-            sum += sinc_table[j as usize] * in_slice[j as usize];
-            j += 1
-        }
-        out_slice[(out_stride * out_sample) as usize] = sum;
-    }
-
-    #[inline(always)]
-    pub fn direct_step_double(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        n: usize,
-        sinc_table: &[f32],
-    ) {
-        let mut accum: [f64; 4] = [0.0; 4];
-        let mut j = 0;
-
-        while j < n {
-            accum[0usize] +=
-                f64::from(sinc_table[j as usize] * in_slice[j as usize]);
-            accum[1usize] += f64::from(
-                sinc_table[(j + 1) as usize] * in_slice[(j + 1) as usize],
-            );
-            accum[2usize] += f64::from(
-                sinc_table[(j + 2) as usize] * in_slice[(j + 2) as usize],
-            );
-            accum[3usize] += f64::from(
-                sinc_table[(j + 3) as usize] * in_slice[(j + 3) as usize],
-            );
-            j += 4
-        }
-        let sum: f64 =
-            accum[0usize] + accum[1usize] + accum[2usize] + accum[3usize];
-        out_slice[(out_stride * out_sample) as usize] = sum as f32;
-    }
-}
-
-#[cfg(target_feature = "sse2")]
-mod sse2 {
-    #[cfg(target_arch = "x86")]
-    use std::arch::x86::*;
-    #[cfg(target_arch = "x86_64")]
-    use std::arch::x86_64::*;
-
-    #[inline]
-    pub unsafe fn convert_m128_to_m256d(v: __m128) -> __m256d {
-        let v64 = _mm256_castpd128_pd256(_mm_cvtps_pd(v));
-        let v64 = _mm256_insertf128_pd(
-            v64,
-            _mm_cvtps_pd(_mm_shuffle_ps(v, v, 0b00001110)),
-            1,
-        );
-        v64
-    }
-
-    #[inline]
-    pub unsafe fn hsum_m256d(v: __m256d) -> f64 {
-        #[cfg(target_arch = "x86")]
-        use std::arch::x86::*;
-        #[cfg(target_arch = "x86_64")]
-        use std::arch::x86_64::*;
-
-        let low = _mm256_extractf128_pd(v, 0);
-        let high = _mm256_extractf128_pd(v, 1);
-        let low = _mm_add_pd(low, high);
-        let high64 = _mm_unpackhi_pd(low, low);
-        _mm_cvtsd_f64(_mm_add_sd(low, high64))
-    }
-
-    #[inline(always)]
-    pub unsafe fn cubic_coef(
-        frac: f32,
-        interp: &mut std::arch::x86_64::__m128,
-    ) {
-        let mut interp_v = [0.0, 0.0, 0.0, 0.0];
-
-        interp_v[0] = -0.16666999459266663 * frac
-            + 0.16666999459266663 * frac * frac * frac;
-        interp_v[1] = frac + 0.5 * frac * frac - 0.5f32 * frac * frac * frac;
-        interp_v[3] = -0.3333300054073334 * frac + 0.5 * frac * frac
-            - 0.16666999459266663 * frac * frac * frac;
-        interp_v[2] = (1.0f64
-            - interp_v[0] as f64
-            - interp_v[1] as f64
-            - interp_v[3] as f64) as f32;
-
-        *interp = _mm_loadu_ps(interp_v.as_ptr());
-    }
-
-    #[inline(always)]
-    pub fn interpolate_step_single(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        oversample: usize,
-        offset: usize,
-        n: usize,
-        sinc_table: &[f32],
-        frac: f32,
-    ) {
-        unsafe {
-            let mut accum = _mm_setzero_ps();
-            in_slice.iter().zip(0..n).for_each(|(&curr_in, j)| {
-                let idx =
-                    (2 + (j + 1) * oversample as usize) - offset as usize;
-                let sinc_ptr: *const f32 = sinc_table[idx..].as_ptr();
-                accum = _mm_add_ps(
-                    accum,
-                    _mm_mul_ps(_mm_loadu_ps(sinc_ptr), _mm_set1_ps(curr_in)),
-                );
-            });
-            let mut interp = _mm_setzero_ps();
-            cubic_coef(frac, &mut interp);
-            let v = _mm_mul_ps(interp, accum);
-            out_slice[(out_stride * out_sample) as usize] =
-                _mm_cvtss_f32(_mm_hadd_ps(_mm_hadd_ps(v, v), v));
-        }
-    }
-
-    pub fn interpolate_step_double(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        oversample: usize,
-        offset: usize,
-        n: usize,
-        sinc_table: &[f32],
-        frac: f32,
-    ) {
-        unsafe {
-            let mut accum = _mm256_setzero_pd();
-            in_slice.iter().zip(0..n).for_each(|(&curr_in, j)| {
-                let idx =
-                    (2 + (j + 1) * oversample as usize) - offset as usize;
-                let sinct_ptr: *const f32 = sinc_table[idx..].as_ptr();
-                let v =
-                    _mm_mul_ps(_mm_loadu_ps(sinct_ptr), _mm_set1_ps(curr_in));
-                let v64 = convert_m128_to_m256d(v);
-                accum = _mm256_add_pd(accum, v64);
-            });
-
-            let mut interp = _mm_setzero_ps();
-            cubic_coef(frac, &mut interp);
-
-            let accum32 = _mm256_cvtpd_ps(accum);
-
-            let v = _mm_mul_ps(interp, accum32);
-            out_slice[(out_stride * out_sample) as usize] =
-                _mm_cvtss_f32(_mm_hadd_ps(_mm_hadd_ps(v, v), v));
-        }
-    }
-
-    #[inline(always)]
-    pub fn direct_step_single(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        n: usize,
-        sinc_table: &[f32],
-    ) {
-        unsafe {
-            let accum = sinc_table
-                .chunks_exact(8)
-                .zip(in_slice.chunks_exact(8))
-                .take((n as usize) / 8)
-                .fold(_mm_setzero_ps(), |acc, (sinct_p, iptr_p)| {
-                    let sinct_v = _mm_loadu_ps(sinct_p.as_ptr());
-                    let iptr_v = _mm_loadu_ps(iptr_p.as_ptr());
-
-                    let acc = _mm_add_ps(acc, _mm_mul_ps(sinct_v, iptr_v));
-
-                    let sinct_v = _mm_loadu_ps(sinct_p[4..].as_ptr());
-                    let iptr_v = _mm_loadu_ps(iptr_p[4..].as_ptr());
-                    _mm_add_ps(acc, _mm_mul_ps(sinct_v, iptr_v))
-                });
-
-            let accum = _mm_add_ps(accum, _mm_movehl_ps(accum, accum));
-            let accum =
-                _mm_add_ss(accum, _mm_shuffle_ps(accum, accum, 0b01010101));
-
-            _mm_store_ss(
-                out_slice[((out_stride * out_sample) as usize)..].as_mut_ptr(),
-                accum,
-            )
-        }
-    }
-
-    #[inline(always)]
-    pub fn direct_step_double(
-        in_slice: &[f32],
-        out_slice: &mut [f32],
-        out_stride: usize,
-        out_sample: usize,
-        n: usize,
-        sinc_table: &[f32],
-    ) {
-        unsafe {
-            let mut accum = _mm256_setzero_pd();
-            let mut j = 0;
-
-            while j < n {
-                let sinct_v = _mm_loadu_ps(sinc_table[j as usize..].as_ptr());
-                let iptr_v = _mm_loadu_ps(in_slice[j as usize..].as_ptr());
-                let v = _mm_mul_ps(sinct_v, iptr_v);
-
-                accum = _mm256_add_pd(accum, convert_m128_to_m256d(v));
-                j += 4;
-            }
-
-            out_slice[(out_stride * out_sample) as usize] =
-                hsum_m256d(accum) as f32;
-        }
-    }
-
-    #[cfg(test)]
-    mod tests {
-        #[cfg(target_arch = "x86")]
-        use std::arch::x86::*;
-        #[cfg(target_arch = "x86_64")]
-        use std::arch::x86_64::*;
-
-        #[test]
-        fn test_convert_m128_to_m256d() {
-            let input: Vec<f32> = vec![1.0, 2.0, 3.0, 4.0];
-            let mut output: Vec<f64> = vec![0.0; 4];
-
-            unsafe {
-                let v = _mm_loadu_ps(input.as_ptr());
-                _mm256_storeu_pd(
-                    output.as_mut_ptr(),
-                    super::convert_m128_to_m256d(v),
-                );
-            }
-
-            assert_eq!(output, vec![1.0, 2.0, 3.0, 4.0]);
-        }
-
-        #[test]
-        fn test_hsum_m256d() {
-            let input: Vec<f64> = vec![1.0, 2.0, 3.0, 4.0];
-
-            let result = unsafe {
-                let v = _mm256_loadu_pd(input.as_ptr());
-                super::hsum_m256d(v)
-            };
-
-            assert_eq!(result as usize, 10 as usize);
-        }
-    }
-}
-
 cfg_if! {
-    if #[cfg(target_feature = "sse2")] {
-        use sse2::*;
+    if #[cfg(feature = "sse3")] {
+        use sse3::*;
     } else {
         use native::*;
     }
@@ -1688,3 +1348,8 @@ fn speex_resampler_magic<'a, 'b>(
     *out = &mut value[(out_len * st.out_stride as u32) as usize..];
     out_len
 }
+
+#[cfg(not(feature = "sse3"))]
+mod native;
+#[cfg(feature = "sse3")]
+mod sse3;
